diagvoice-backend/requirements.txt
torch
torchaudio
transformers
safetensors
fastapi
uvicorn
flask
pyyaml
sentence-transformers
python-multipart
librosa
pydub
accelerate
soundfile
opensmile
parselmouth
openpyxl
pandas
scikit-learn
joblib
matplotlib
seaborn
